Removed interface PermissionTemplateDAO
